requirements.txt
pandas
requests
python-dotenv
fastapi
uvicorn
pydantic
tabulate
tenacity
openai
